refactor(python): separate x response inspection
